A lot of changes Moved on Ogre 12 Implemented Compound (AKA advanced JSON) Implemented conversion interfaces and conversion functionality between types Implemented Serialization interface and it's functionality in some classes Reworked database functionality, Table is used now and data may be stored in json files Started converting more asset types to work with ogre

+++ Engine/File.cpp
@@ -112,7 +112,7 @@ uint File::Reader::get_size() const
 
 uint File::Reader::get_position()
 {
-	return static_cast<uint>(stream_.tellg());
+	return stream_.tellg();
 }
 
 Shared<File::Writer> File::Writer::open(const Path& path)
